Split event and base loaded tests
