+++ tests/event/test_event.py
@@ -1,42 +1,8 @@
-import json
-from unittest.mock import patch
-
-import pytest
-
-from corva.event.event import Event
 from corva.event.data.base import BaseEventData
+from corva.event.event import Event
 
 
-def test_load_empty_event():
-    event = ''
-    with pytest.raises(ValueError) as exc:
-        Event._load(event=event)
-    assert str(exc.value) == 'Empty event'
-
-
-def test_load_wrong_event_type():
-    event = {'': ''}
-    with pytest.raises(TypeError) as exc:
-        Event._load(event=event)
-    assert str(exc.value) == f'Unknown event type {type(event)}'
-
-
-def test_load_invalid_json():
-    event = '{}'
-    with patch('corva.event.base.json.loads', side_effect=ValueError):
-        with pytest.raises(ValueError) as exc:
-            Event._load(event=event)
-    assert str(exc.value) == 'Invalid event JSON'
-
-
-def test_load():
-    event = {'key1': 'val1'}
-
-    loaded = Event._load(event=json.dumps(event))
-    assert loaded == event
-
-
-def test_iter(patch_base_event):
+def test_iter():
     data1 = BaseEventData()
     data2 = BaseEventData(a=1)
     e2 = Event(data=[data1, data2])
@@ -44,7 +10,7 @@ def test_iter(patch_base_event):
         assert e2.data[idx] == data
 
 
-def test_eq(patch_base_event):
+def test_eq():
     class CustomEventData(BaseEventData):
         pass
 
@@ -74,13 +40,7 @@ def test_eq(patch_base_event):
     assert e1 == e2
 
 
-def test_len(patch_base_event):
+def test_len():
     assert len(Event(data=[])) == 0
     assert len(Event(data=[BaseEventData()])) == 1
     assert len(Event(data=[BaseEventData(), BaseEventData()])) == 2
-
-
-def test_abstractmethods():
-    assert getattr(Event, '__abstractmethods__') == frozenset(['load'])
-    with pytest.raises(TypeError):
-        Event(data=[])
